add -Wnarrowing as error to cmake builds (#89207)
Pull Request resolved: https://github.com/pytorch/pytorch/pull/89207
Approved by: https://github.com/wconstab, https://github.com/malfet

+++ aten/src/ATen/native/mps/operations/Distributions.mm
@@ -438,7 +438,7 @@ Tensor& multinomial_with_replacement_mps_kernel(
           MPSGraphTensor *randomTensor = generatorTensors[0];
 
           auto broadcastShape = @[ns_numDist ,ns_n_sample, ns_numCategories];
-          int broadcastShapeVals[3] = {numDist, n_sample, numCategories};
+          int broadcastShapeVals[3] = {numDist, static_cast<int>(n_sample), numCategories};
           MPSGraphTensor *broadcastShapeTensor = [mpsGraph constantWithData:[NSData dataWithBytes:broadcastShapeVals length:sizeof(int) * broadcastShape.count]
                                                                       shape:@[[NSNumber numberWithUnsignedInteger:broadcastShape.count]]
                                                                    dataType:MPSDataTypeUInt32];
